some fixes

+++ modules/benchmark-maprfs-rwtest/src/main/groovy/com/mapr/emea/ps/clustercheck/module/benchmark/maprfs/rwtest/BenchmarkMaprFsRwTestModule.groovy
@@ -231,8 +231,8 @@ sleep 3
                             compression: compression,
                             dataSizeInMB: sizeInMB as Long,
                             threads: threads as Long,
-                            writeRatesInMBperSecond: writeRates as Long,
-                            readRatesInMBperSecond: readRates as Long,
+                            writeRatesInMBperSecond: writeRates,
+                            readRatesInMBperSecond: readRates,
                             sumWriteRateInMBperSecond: writeRates.sum(),
                             sumReadRateInMBperSecond: readRates.sum()
                     ]
